move contraintsbuilder to model
